Chatbot/backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
pyjwt==2.9.0
bcrypt==4.2.1
chromadb==0.5.23
sentence-transformers==3.4.1
scikit-learn==1.6.1
numpy==2.2.3
torch==2.6.0
google-generativeai==0.8.4
python-dotenv==1.0.1
